services/schedule_service/requirements.txt
# FastAPI и веб-сервер
fastapi==0.115.9
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# База данных
asyncpg==0.30.0
SQLAlchemy==2.0.38
alembic==1.14.1

# Redis для кэширования
redis==5.0.1

# Безопасность
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Валидация и конфигурация
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1

# HTTP клиент для межсервисного взаимодействия
httpx==0.28.1

# Утилиты
pytz==2024.2
